% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Translate strings}
\usage{
tr(id, language = getOption("translatr.language"))
}
\arguments{
\item{id}{the key referencing the desired text}

\item{language}{the language to provide. This is best set using \code{options(translatr.language = "language")}.}
}
\value{
a translated character string
}
\description{
This function uses i18n translations files to provide multi-lingual support for GUIs build with R.
}
\examples{
cat('Please see README')
}
