% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{tr}
\alias{tr}
\title{Translate text}
\usage{
tr(textid, language, table, id_col = 1L, con, tbl)
}
\arguments{
\item{textid}{the character ID of the string to translate}

\item{language}{the language code to translate to}

\item{table}{a data.frame-like object with translations}

\item{id_col}{index or name of the column containing text IDs}

\item{con}{a database connection containing necessary translation tables}

\item{tbl}{the name of the table in the database containing the translations}
}
\description{
The translate function returns the translation of a text by ID.
}
\details{
You can set the default language for your session using \code{options(translatr.language = 'mri')}, otherwise "en" will be used (by default). Similarly, you can specify the translation dictionary with the option \code{translatr.table}, or by supplying \code{translatr.con} and \code{translatr.tbl} to connect to a database with a table containing the translations.
}
\examples{
trans_df <- data.frame(id = "welcome", en = "Welcome", mri = "Haere mai")
tr("welcome", "mri", trans_df)

\dontrun{
require(RSQLite)
con <- dbConnect(SQLite(), system.file("introduction.sqlite", package = "translatr"))
options(translatr.con = con, translatr.tbl = "intro")
tr("welcome", "fr")
}
}
